AP_Mount: support 32 bit system IDs
Widen gimbal and target sysid storage and pass 32 bit targets via the
extended header aware send path. MAV_CMD_DO_GIMBAL_MANAGER_CONFIGURE
accepts a controller sysid up to 2^24-1, the largest a float command
parameter can carry exactly.

+++ libraries/AP_Mount/AP_Mount_MAVLink.h
@@ -98,7 +98,7 @@ private:
     bool _initialised;              // true once the gimbal has provided a GIMBAL_DEVICE_INFORMATION
     uint32_t _last_devinfo_req_ms;  // system time that GIMBAL_DEVICE_INFORMATION was last requested (used to throttle requests)
     class GCS_MAVLINK *_link;       // link we have found gimbal on; nullptr if not seen yet
-    uint8_t _sysid;                 // sysid of gimbal
+    uint32_t _sysid;                // sysid of gimbal
     uint8_t _compid;                // component id of gimbal
     mavlink_gimbal_device_attitude_status_t _gimbal_device_attitude_status;  // copy of most recently received gimbal status
     uint32_t _last_attitude_status_ms;  // system time last attitude status was received (used for health reporting)
